Package 'sys' renamed to 'util'.
SERVER_HOST changed to '0.0.0.0' as '127.0.0.1' was only accepting inbound TFTP connections via the loopback interface.  Using '0.0.0.0' as the host binds to any/all network interfaces.

SERVER_PORT changed to well known TFTP port number 69.  Run node as Administrator (Windows) or root (*nix) to be allowed to bind to this port.